editing/cefalse-boundaries-deletion-manual.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Selecting and deleting all from the cE=true element with cE=false element at boundaries</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  #host {
    border: 2px red solid;
  }
</style>
<div contenteditable id="host">
  <div contenteditable="false" id="cefalsebeginning">&nbsp;</div>
  <p id="paragraph">Lorem ipsum dolor sit amet.</p>
  <div contenteditable="false" id="cefalseend">&nbsp;</div>
</div>

<ol>
  <li>Select all from the red area above.</li>
  <li>Delete it by pressing <kbd>Backspace</kbd>/<kbd>Delete</kbd>.</li>
</ol>
<script>
  setup( { explicit_timeout: true } );

    promise_test( () => {
    const cefalsebeginning = document.querySelector( '#cefalsebeginning' );
    const cefalseend = document.querySelector( '#cefalseend' );
    const paragraph = document.querySelector( '#paragraph' );
    return new Promise( ( resolve, reject ) => {
        host.addEventListener( 'keyup', ( { key }) => {
          if ( key !== 'Backspace' && key !== 'Delete' ) {
            return;
          }

          try {
              assert_false( cefalsebeginning.isConnected, 'cE=false element was removed' );
              assert_false( cefalseend.isConnected, 'cE=false element was removed' );
              assert_false( paragraph.isConnected, 'paragraph was removed' );
          } catch ( error ) {
              reject( error );
          }
          resolve();
        } );
    } );
    } );
</script>
